# openapi/certifications-v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Certifications.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/employees/{employeeId}/contracts/current": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get an employee's active contract",
        "operationId": "GetCurrentEmployeeContract",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/employees/{employeeId}/contracts": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Create a new active contract after the previous contract is closed",
        "operationId": "CreateEmployeeContract",
        "parameters": [
          {
            "name": "employeeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/contracts/{contractId}/close": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Close an active contract",
        "operationId": "CloseContract",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseContractRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
    "/api/v1/me/contract": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get the current employee's active contract and certification history",
        "operationId": "GetOwnContract",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ],
            "CookieAuth": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CertificationDto": {
        "type": "object",
        "properties": {
          "certificationId": {
            "type": "integer",
            "format": "int64"
          },
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "assessor": {
            "type": "string",
            "nullable": true
          },
          "certificationDate": {
            "type": "string",
            "format": "date"
          },
          "protocolDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationSend": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationReturned": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "isCompleted": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "CertificationStatus": {
        "enum": [
          0,
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "CloseContractRequest": {
        "type": "object",
        "properties": {
          "closedOn": {
            "type": "string",
            "format": "date"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "ContractDetailsDto": {
        "type": "object",
        "properties": {
          "contract": {
            "$ref": "#/components/schemas/ContractDto"
          },
          "certifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CertificationDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ContractDto": {
        "type": "object",
        "properties": {
          "contractId": {
            "type": "integer",
            "format": "int64"
          },
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "effectiveValidTo": {
            "type": "string",
            "format": "date"
          },
          "active": {
            "type": "boolean"
          },
          "prolongationWarningMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationAlertMonths": {
            "type": "integer",
            "format": "int32"
          },
          "prolongationForYears": {
            "type": "integer",
            "format": "int32"
          },
          "rowVersion": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "$ref": "#/components/schemas/CertificationStatus"
          }
        },
        "additionalProperties": false
      },
...
      "CreateContractRequest": {
        "type": "object",
        "properties": {
          "position": {
            "type": "string",
            "nullable": true
          },
          "department": {
            "type": "string",
            "nullable": true
          },
          "division": {
            "type": "string",
            "nullable": true
          },
          "contractDate": {
            "type": "string",
            "format": "date"
          },
          "validTo": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "prolongationWarningMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "prolongationAlertMonths": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "prolongationForYears": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Additional API-channel key required by every /api/v1 operation.",
        "name": "X-API-Key",
        "in": "header"
      },
      "CookieAuth": {
        "type": "apiKey",
        "description": "Secure HttpOnly authentication cookie issued by Login.",
        "name": "Certifications.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}